Fix scolling to first post via `m.route.set`
The default first post number is '1', so we scroll to that if we're calling `m.route.set` without a `near` parameter, as that means we're scrolling to the top.

This was present in beta 13's implementation, but accidentially omitted in 3f0b3c7.

We also remove unnecessary typecasting for simpler logic and increased consistency with beta 13.